+++ CHANGELOG.md
@@ -11,6 +11,16 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.0.0-rc.??? (2021-??-??)
+
+#### :rocket: New Feature
+
+* Now a route pattern can be a function
+
+#### :house: Internal
+
+* Extracted `compileStaticRoutes` from `bRouter` to `core/router`
+
 ## v3.0.0-rc.181 (2021-04-20)
 
 #### :bug: Bug Fix